% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bayesest_ex}
\alias{bayesest_ex}
\title{estimate a network formation model using the approximate exchange algorithm}
\usage{
bayesest_ex(g, x, a0, ppp, dt, paramsample, netsample, covpars, mu0, sigma0,
  skip = 1, pr = 0, pc = 0, pf = 0, pinv = 0, pb = 0, nmodes = 1,
  modes = 0.5, sizelargestep = 10, mtemp = 2, seed = 1977)
}
\arguments{
\item{g}{the matrix containing the network}

\item{x}{the matrix of exogenous variables}

\item{a0}{parameter vector at which the simulation is started}

\item{ppp}{vector of indicators for the utility functions}

\item{dt}{indicators of which statistics to compute}

\item{netsample}{number of networks to sample}

\item{covpars}{matrix of covariances among parameters (for sampler)}

\item{mu0}{vector of means for the prior}

\item{sigma0}{vector of variances for the prior}

\item{skip}{how many networks to skip between consecutive sample}

\item{pr}{probability of changing a row}

\item{pc}{probability of changing a column}

\item{pf}{probability of large step}

\item{pinv}{probability of inversion of adjacency matrix}

\item{pb}{probability of updating using Bhamidi-like step
(default \code{pb = 0})}

\item{nmodes}{number of modes of gibbs distribution
(default \code{nmodes = 1})}

\item{modes}{modes of the gibbs distribution
(default \code{modes = 0.5})}

\item{sizelargestep}{how many links to update in a large step}

\item{mtemp}{number of temperatures for simulated tempering (not implemented)}

\item{seed}{sets the seed for the random number generator (an integer)}

\item{parsample}{number of parameters to sample}
}
\description{
This function estimates the posterior of structural parameters of
a network formation model, using the approximate exchange algorithm
used in Mele (2015). For each proposed parameter, it simulates
networks from the stationary distribution
starting at network \code{g}, and saves the sample in data.frame
\code{parsample}: each column is a parameter associated with a
network statistics. The
algorithm can perform small steps (1 links update per iteration),
intermediate steps (1 row or 1 column update per iteration),
or large step (update a number of links determined by the control
variable \code{sizelargestep}). If the model is an homogenous
player model, we can also use the Bhamidi et al (2011) approach
to establish the modes of the gibbs distribution and use that
in the simulation, i.e. we propose to jump to one of the
modes with probability \code{pb}.
}
\examples{
dt  <- matrix(data = c(1,1,1,1,1,1), nrow = 3, ncol = 2)
p   <- c(1, 2, 3)
x   <- matrix(data = rnorm(5,5,1), nrow = 5, ncol = 1)
g   <- matrix(data = rbinom(25,1,.5), nrow = 5, ncol = 5)
a0   <- c(-4,-1,1)
mu0  <- c(0,0,0)
sigma0 <- c(1,1,1)
covpars <- matrix(c(1,0,0,0,1,0,0,0,1), ncol = 3, nrow = 3)
ans <- bayesest_ex(g, x, a0, p, dt, parsample = 100, netsample = 100
                     mu0 = mu0 , sigma0 = sigma0, covpars = covpars)
ans
}
\author{
Angelo Mele
}

