% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bhamidiedgetriangle}
\alias{bhamidiedgetriangle}
\title{computes solutions to equation in Bhamidi et al 2011 to check for high/low temp phase (edge - triangle model)}
\usage{
bhamidiedgetriangle(a, b)
}
\arguments{
\item{a}{parameter for edges term}

\item{b}{parameter for cyclic triangles term}
}
\description{
this function computes solutions to equation in Bhamidi et al 2011 for the directed network.
Returns a matrix with solution (row 1) and first derivative at the solution (row 2)
}
\examples{
bhamidiedgetriangle(a = -2, b = 0.2)
}
\author{
Angelo Mele
}

