% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bhamidieit}
\alias{bhamidieit}
\title{computes solutions to equation in Bhamidi et al 2011 to check for high/low temp phase  (edges + indirect + triangles model)}
\usage{
bhamidieit(a, b, c)
}
\arguments{
\item{a}{parameter for edges term}

\item{b}{parameter for reciprocity term}
}
\description{
this function computes solutions to equation in Bhamidi et al 2011 for the directed network.
Returns a matrix with solution (row 1) and first derivative at the solution (row 2)
}
\examples{
bhamidieit(a = -2, b = 0.2, c = 0.1)
}
\author{
Angelo Mele
}

