% Generated by roxygen2 (4.0.2): do not edit by hand
\name{erdos}
\alias{erdos}
\title{simulate a sample from directed erdos-renyi model}
\usage{
erdos(p, n)
}
\arguments{
\item{p}{probability of a link}

\item{n}{number of nodes of the network}
}
\description{
This function simulates a network from the erdos-renyi model with
probability of link \code{p} and \code{n} nodes. The result is
an adjacency matrix
}
\examples{
g <- erdos(p = 0.5, n = 10)
g
}
\author{
Angelo Mele
}

