% Generated by roxygen2 (4.0.2): do not edit by hand
\name{netstats}
\alias{netstats}
\title{Computes network statistics}
\usage{
netstats(ppp, ggg, xxx, dtt)
}
\arguments{
\item{ppp}{vector of indicators for the utility functions}

\item{ggg}{the matrix containing the network}

\item{xxx}{the matrix of exogenous variables}

\item{dtt}{indicators of which statistics to compute}
}
\description{
This function computes all the network statistics for network \code{g}
}

