% Generated by roxygen2 (4.0.2): do not edit by hand
\name{phi_edgetriangle}
\alias{phi_edgetriangle}
\title{function from Bhamidi et al 2011 to check for high/low temp phase (edge - triangle model)}
\usage{
phi_edgetriangle(p, a, b)
}
\arguments{
\item{p}{prob of link (mean field approximation)}

\item{a}{parameter for edges term}

\item{b}{parameter for cyclic triangle term}
}
\description{
this function is the same as Bhamidi et al 2011 for the directed network
}
\examples{
phi_edgetriangle(p = 0.2, a = -2, b = 0.2)
}
\author{
Angelo Mele
}

