% Generated by roxygen2 (4.0.2): do not edit by hand
\name{samplemany}
\alias{samplemany}
\title{simulate a sample of networks from stationary distribution}
\usage{
samplemany(g, x, a, ppp, dt, samplesize, skip = 1, pr = 0, pc = 0,
  mtemp = 2, seed = 1977)
}
\arguments{
\item{g}{the matrix containing the network}

\item{x}{the matrix of exogenous variables}

\item{ppp}{vector of indicators for the utility functions}

\item{dt}{indicators of which statistics to compute}

\item{samplesize}{length of the sample}

\item{skip}{how many networks to skip between consecutive sample}

\item{pr}{probability of changing a row}

\item{pc}{probability of changing a column}

\item{mtemp}{number of temperatures for simulated tempering (not implemented)}

\item{seed}{sets the seed for the random number generator (an integer)}
}
\description{
This function simulates networks from the stationary distribution
starting at network \code{g}, and saves the sample in data.frame
\code{netsample}: each column is a network statistics.
}
\examples{
dt  <- matrix(data = c(1,1,1,1,1,1), nrow = 3, ncol = 2)
p   <- c(1, 2, 3)
x   <- matrix(data = rnorm(5,5,1), nrow = 5, ncol = 1)
g   <- matrix(data = rbinom(25,1,.5), nrow = 5, ncol = 5)
a   <- c(-4,-1,1)
ans <- samplemany(g, x, a, p, dt, skip=100, samplesize = 1000)
ans
}
\author{
Angelo Mele
}

