% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simulatemodellarge}
\alias{simulatemodellarge}
\title{simulate selected models with few parameters}
\usage{
simulatemodellarge(model = "erdos-renyi", initg, a, samplesize = 100,
  skip = 10, pr = 0, pc = 0, pf = 0, sizelargestep = 10, mtemp = 2,
  seed = 1977)
}
\arguments{
\item{model}{a string indicating which model you want to
simulate. The default is \code{"erdos-renyi"}. Other choices implemented are
a model with mutual links (\code{"recip"}) and one with
3 parameters (\code{"3params"})}

\item{initg}{the initial network for the simulations}

\item{a}{the vector of parameters}

\item{samplesize}{the number of simulated networks to collect (default is
\code{samplesize =  100})}

\item{skip}{how many networks to skip between sampled networks, with
default \code{skip = 10}.}

\item{pr}{probability of swapping a row of network matrix
(default \code{pr = 0})}

\item{pc}{probability of swapping a column of network matrix
(default \code{pc = 0})}

\item{pf}{probability of updating a large number of links
(default \code{pf = 0})}

\item{sizelargestep}{number of links to update in large step
(default \code{sizelargestep = 10})}

\item{mtemp}{number of temperatures for simulated tempering (not implemented)}

\item{seed}{seed for random number generator}
}
\description{
runs a simulation of simple models (erdos-renyi, mutual links)
default is erdos-renyi model
}
\examples{
n  <- 100
dt <- matrix(data = c(1,1,1,1,1,1), nrow = 3, ncol = 2)
p  <- c(1, 2, 3)
x  <- matrix(data = rnorm(n,5,1), nrow = n, ncol = 1)
g  <- matrix(data = rbinom(n*n,1,.5), nrow = n, ncol = n)
a  <- c(-2,.5,.01)
# show data
p; g; x; dt
tt <- simulatemodellarge("3params", g, a, skip = 1000, samplesize = 10000)
tt
}
\seealso{
\link[net]{samplemany}
}

